requirements.txt
tensorflow==2.9.1     
tqdm
streamlit
pillow
numpy



